ImageStyleClassifier/requirements.txt
gradio==5.9.1
tensorflow==2.17.0
numpy==1.26.4
Pillow==10.4.0

